Modify some codes to support parallel calculations with MPI for 2D rectangle domain.

+++ FElib/src/mesh/scale_meshutil_2d.F90
@@ -459,7 +459,7 @@ contains
     !-----------------------------------------------------------------------------
 
     NtilePerPanel = Ntile/1
-    NeY = sqrt(dble(NtilePerPanel))
+    NeY = int( sqrt(dble(NtilePerPanel)) )
     NeX = NtilePerPanel/NeY
     NvX = NeX + 1
     NvY = NeY + 1
@@ -473,7 +473,6 @@ contains
       nodesID_2d(i,j) = counter
     end do
     end do
-
 
     !----
